+++ tests/test_reorder_offsets.c
@@ -24,7 +24,7 @@ int32_t *data;
 int32_t *data_dest;
 
 static char* test_reorder_offsets(void) {
-  size_t isize = CHUNKSIZE * sizeof(int32_t);
+  int32_t isize = CHUNKSIZE * sizeof(int32_t);
   int dsize;
   blosc2_cparams cparams = BLOSC2_CPARAMS_DEFAULTS;
   blosc2_dparams dparams = BLOSC2_DPARAMS_DEFAULTS;
